pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reroom"
version = "0.1.0"
description = "Reimage your Room via AI — transfer design concepts between room photographs."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "ReRoom Contributors"}]
keywords = ["ai", "interior-design", "room-design", "azure", "gpt-4o", "image-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "azure-ai-inference>=1.0.0b1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.13.0",
]

[project.scripts]
reroom = "reroom.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
